basic implementation needs calendar interval parsing
