+++ standalone/src/main/scala/io/delta/standalone/internal/exception/DeltaErrors.scala
@@ -273,6 +273,15 @@ private[internal] object DeltaErrors {
       s"Parsed $nestType type:\n${nested.toPrettyJson}")
   }
 
+  def cannotModifyTableProperty(prop: String): Throwable =
+    throw new UnsupportedOperationException(
+      s"The Delta table configuration $prop cannot be specified by the user")
+
+  def unknownConfigurationKeyException(confKey: String): Throwable = {
+    // TODO: AnalysisException ?
+    new IllegalArgumentException(s"Unknown configuration was specified: $confKey")
+  }
+
   ///////////////////////////////////////////////////////////////////////////
   // Helper Methods
   ///////////////////////////////////////////////////////////////////////////
